+++ packages/graphql-connection-transformer/src/ModelConnectionTransformer.ts
@@ -18,6 +18,7 @@ import {
   SortKeyFieldInfoTypeName,
   CONDITIONS_MINIMUM_VERSION,
   makeAttributeTypeEnum,
+  makeEnumFilterInputObjects,
 } from 'graphql-dynamodb-transformer';
 import {
   getBaseType,
@@ -42,6 +43,7 @@ const CONNECTION_STACK_NAME = 'ConnectionStack';
 interface RelationArguments {
   keyName?: string;
   fields: string[];
+  limit?: number;
 }
 
 function makeConnectionAttributeName(type: string, field?: string) {
@@ -611,6 +613,7 @@ export class ModelConnectionTransformer extends Transformer {
         args.fields,
         keySchema,
         index ? String(index.IndexName) : undefined,
+        args.limit,
       );
 
       ctx.setResource(ResolverResourceIDs.ResolverResourceID(parentTypeName, fieldName), queryResolver);
@@ -673,6 +676,14 @@ export class ModelConnectionTransformer extends Transformer {
   ): void {
     const scalarFilters = makeScalarFilterInputs(this.supportsConditions(ctx));
     for (const filter of scalarFilters) {
+      if (!this.typeExist(filter.name.value, ctx)) {
+        ctx.addInput(filter);
+      }
+    }
+
+    // Create the Enum filters
+    const enumFilters = makeEnumFilterInputObjects(field, ctx, this.supportsConditions(ctx));
+    for (const filter of enumFilters) {
       if (!this.typeExist(filter.name.value, ctx)) {
         ctx.addInput(filter);
       }