+++ packages/amplify-category-interactions/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.3.7...amplify-category-interactions@2.3.8) (2020-05-26)
+
+**Note:** Version bump only for package amplify-category-interactions
+
+
+
+
+
+## [2.3.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.3.6...amplify-category-interactions@2.3.7) (2020-05-15)
+
+**Note:** Version bump only for package amplify-category-interactions
+
+
+
+
+
 ## [2.3.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-interactions@2.3.5...amplify-category-interactions@2.3.6) (2020-05-08)
 
 **Note:** Version bump only for package amplify-category-interactions