+++ packages/amplify-category-predictions/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.3.8...amplify-category-predictions@2.3.9) (2020-05-26)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
+## [2.3.8](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.3.7...amplify-category-predictions@2.3.8) (2020-05-15)
+
+**Note:** Version bump only for package amplify-category-predictions
+
+
+
+
+
 ## [2.3.7](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-predictions@2.3.6...amplify-category-predictions@2.3.7) (2020-05-08)