+++ packages/amplify-category-storage/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.6](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-storage@2.4.5...amplify-category-storage@2.4.6) (2020-05-26)
+
+**Note:** Version bump only for package amplify-category-storage
+
+
+
+
+
+## [2.4.5](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-storage@2.4.4...amplify-category-storage@2.4.5) (2020-05-15)
+
+**Note:** Version bump only for package amplify-category-storage
+
+
+
+
+
 ## [2.4.4](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-storage@2.4.3...amplify-category-storage@2.4.4) (2020-05-08)
 
 **Note:** Version bump only for package amplify-category-storage